drop file type filter when picking a directory
Previously, the file filter was set to "*.*" which
somehow led to all directories being disabled as
not clickable / selectable in the directory picker
dialog on Linux (Ubuntu 19). As stated in the docs
of "gen2brain/dlgs", setting the file filter to a
blank string will display all file types.
Tested on Ubuntu 19. Works.