feat: pass multiple types of events from channel
Need to use event channel to communicate between native code and dart
but existing implementation expected only a boolean value to be passed.

This creates an file event which translates to a platform event. The
setup allows for future work related to passing more events. This commit
includes the addition of the source paths as soon as they are available
as they may be usable more immediately than the copied files are in the
event of large files.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>