Updated Dependency Win32 3.0.0
As mentioned in Win32 Change Log,

"COM and Windows Runtime methods and properties are now camelCased, not TitleCased. This is inconvenient, but it avoids a whole class of name clashes and aligns COM and WinRT APIs more closely with Dart idioms. As the projections get smarter with more helpers, we think this is the right call for the future and worth a one-time tax to fix."

Test Passed: test\file_picker_windows_test.dart