saveFile() method for android platform interface changed
now the saveFile method for android uses the fileName as
input file from where it reads the content to be written
in the file that the user is going to choose. this way
there is no need to ask for permission and no need to
delete the 0 byte file created by the save filePicker
window. From the android app ONLY you need to write
the file in a temporary file before calling the saveFile()
method. then you pass the temp file to the saveMethod(),
the user choose the file where the content of the temp
file will be saved and the temp file is written to the
file chosen by the user